EOS-24415: Update dir structure and startup endpoint scripts.
