+++ logrotation/test/py-utils/python_requirements.ext.txt
@@ -1,0 +1,7 @@
+#Python packages transitional dependencies
+# depdendencies of cryptography
+cffi==1.14.5
+# depdendencies of python==3.6
+numpy==1.19.2
+# dependency of cifi
+pycparser==2.20
