+++ logrotation/test/py-utils/python_requirements.txt
@@ -1,0 +1,34 @@
+aiohttp==3.6.1
+argparse==1.4.0
+boto3==1.10.27
+configparser==4.0.2
+confluent-kafka==1.5.0
+cryptography==2.8
+dataclasses==0.7
+dict2xml==1.6.1
+elasticsearch==7.12.0
+elasticsearch-dsl==7.3.0
+Flask==1.1.1
+jsonschema==3.2.0
+marshmallow==3.2.1
+matplotlib==3.1.3
+netifaces==0.10.9
+networkx==2.4
+paramiko==2.7.1
+pathlib==1.0.1
+pika==1.1.0
+prettytable==0.7.2
+pyinotify==0.9.6
+python-consul==1.1.0
+python-crontab==2.5.1
+python-daemon==2.2.4
+PyYAML==5.1.2
+requests==2.25.1
+schematics==2.1.0
+secure==0.2.1
+statsd==3.2.2
+toml==0.10.0
+xmltodict==0.12.0
+zope.component==4.6.2
+zope.event==4.5.0
+zope.interface==5.2.0
